requirements.txt
python-dotenv
numpy
requests
z3
